Extract legacy apply pipeline
